-Fix up installers ability to work out what runtimes are required -revert GetTickCount to keep XP support -switch to building libpng 1.7

+++ Tuniac1/SVPVisualPlugin/SVPRenderer.h
@@ -57,7 +57,7 @@ protected:
 	RECT								m_NextVisRect;
 	RECT								m_PrevVisRect;
 
-	unsigned long long					m_LastMove;
+	unsigned long						m_LastMove;
 	int									m_iElaspedTime;
 
     BITMAP								m_ArrowBM;
